+++ learningCodableJSON/DicWeather.swift
@@ -1,0 +1,14 @@
+//
+//  DicWeather.swift
+//  learningCodableJSON
+//
+//  Created by Larissa Diniz  on 13/02/20.
+//  Copyright © 2020 Larissa Diniz . All rights reserved.
+//
+
+import Foundation
+
+struct DicWeather: Codable {
+    
+    let weather: [Weather]
+}
