+++ learningCodableJSON/ViewController.swift
@@ -17,8 +17,7 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-            
-            
+    
     }
 }
